Update DecisionTree.java
Added all the methods to build the tree. Added a predict() method to take input and do the prediction based on the built tree. New datatypes added to keep track of the encoded columns, column names and the number of features. Keeping track of the encoded columns will help encode the input at the time of prediction. Name of columns are required to know whether there is a "target" column in the dataframe or not. Number of features are matched at the time of prediction to check whether the input on which the prediction is to be performed has all the features or not.
With this, it is a complete, working DecisionTree Classifier class in java.